refactor(guard): #1218 Step 1a — extract policy engine to app/guard/policy.py
First module of the three-part decomposition (policy · audit · router).

Moves from apps/api/app/modules/guard/routers/proxy.py to
apps/api/app/guard/policy.py (verbatim, behavior-identical):
- evaluate()          — was _evaluate_policies (public API, module-level)
- flatten_prompt()    — was _flatten_prompt (public — used by proxy audit-summary too)
- _is_proxy_rule      — helper
- _rule_matches       — helper
- _defense_score      — helper
- SEVERITY_WEIGHTS    — public constant
- _ACTION_RANK        — module constant

proxy.py keeps compat re-exports under the original private names so no
external caller breaks (workflows.py, tests). Zero behavior change.

Regression harness: 8/8 green against docker Postgres.
Downstream: 239 passed (all guard_proxy / guard_pack_matrix / guard_layered_verdict).

One test update — test_guard_layered_verdict.py monkeypatched
compute_policy/set_workspace_rls on proxy_module. Point it at the new
source-of-truth module (app.guard.policy) instead of the re-export layer.
Symbols are unchanged.

Next: Step 1b — extract audit sink to app/guard/audit.py.
